+++ contracts/mocks/Create2Impl.sol
@@ -6,11 +6,11 @@ import "../utils/Create2.sol";
 import "../introspection/ERC1820Implementer.sol";
 
 contract Create2Impl {
-    function deploy(uint value, bytes32 salt, bytes memory code) public {
+    function deploy(uint256 value, bytes32 salt, bytes memory code) public {
         Create2.deploy(value, salt, code);
     }
 
-    function deployERC1820Implementer(uint value, bytes32 salt) public {
+    function deployERC1820Implementer(uint256 value, bytes32 salt) public {
         // solhint-disable-next-line indent
         Create2.deploy(value, salt, type(ERC1820Implementer).creationCode);
     }
